Moved UserIndicator to containers
